Phase-II-Full-Stack-Web-Application/backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.20
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.30.0
alembic==1.13.0
pytest==8.3.3
httpx==0.27.0
python-dotenv==1.0.1